+++ test.py
@@ -39,6 +39,13 @@ def rainbowCycle(strip, wait_ms=2, iterations=5):
 		strip.show()
 		time.sleep(wait_ms/1000.0)
 
+def colorWipe(strip, color, wait_ms=50):
+    """Wipe color across display a pixel at a time."""
+    for i in range(strip.numPixels()):
+        strip.setPixelColor(i, color)
+        strip.show()
+        time.sleep(wait_ms/1000.0)
+
 
 # Main program logic follows:
 if __name__ == '__main__':
@@ -49,5 +56,6 @@ if __name__ == '__main__':
 	strip.begin()
 	print ('Rainbow animations.')
 	rainbowCycle(strip)
+	colorWipe(strip, Color(0,0,0), 10)
 
 quit()